+++ changelog.md
@@ -6,6 +6,7 @@ Codion Change Log
 - SingleSelectionModel and SelectionModel renamed SingleSelection and Selection respectively.
 - FilterModel.clear() moved to Items.
 - FilterModel.Refresher.refresh() removed, refresh() now protected in AbstractRefresher.
+- ColumnSummaryModel renamed SummaryModel.
 ### is.codion.swing.common.model
 - FilterComboBoxModel.includeNull() and nullItem() moved to NullItem.
 - FilterComboBoxModel.replace() moved to FilterComboBoxModelItems.